Handle multiple input values
we look for a comma separated list in the query parameter data.
eg:
    ?data=1,2,3,4,5
Single vales (eg ?data=5) are still fine.

The output format has changed from just the single result value
to a list of records.
example (quotes omitted from json output):
    input: ?data=1,2,3
    output: [{in: 1, out: 1}, {in: 2, out: 2}, {in: 3, out: Fizz}]

We need to return the input values because any inputs that cannot
be converted to ints are silently filtered out.

This output format was chosen over a plain old mapping (something
like: {1: 1, 2: 2, 3: Fizz}) because it should be a lot easier to
work with, and we're all about Developer Experience here at
FizzBuzzaaS.